moved NSP method
